test_match/requirements.txt
winput>=1.4.0
Pillow
pyautogui
numpy